refactor(local): organize model-selection into per-provider slices
Move the Ollama pull UI (catalog, hook, ndjson transport) under providers/ollama/ and split its catalog/pull client out of the shared api.ts, so each provider owns a self-contained vertical slice.
