+++ surfsense_local/frontend/src/features/model-selection/providers/ollama/pull-stream.test.ts
